whatsapp-remindbot/requirements.txt
flask==3.0.3
groq==0.11.0
apscheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
